Extract enums to own file
