+++ TeslaLib/Models/DriverSide.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.Serialization;
+
+namespace TeslaLib.Models
+{
+    public enum DriverSide
+    {
+        [EnumMember(Value = "LH")]
+        LEFT_HAND_DRIVE,
+
+        [EnumMember(Value = "RH")]
+        RIGHT_HAND_DRIVE,
+    }
+}
