+++ TeslaLib/Models/InteriorDecor.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace TeslaLib.Models
+{
+    public class InteriorDeco
+    {
+        public InteriorDeco()
+        {
+        }
+    }
+}
