+++ TeslaLib/Models/Model.cs
@@ -1,0 +1,10 @@
+﻿namespace TeslaLib.Models
+{
+    public enum Model
+    {
+        Unknown,
+        S,
+        X,
+        Three
+    }
+}
