+++ TeslaLib/Models/PanoramicRoofState.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace TeslaLib.Models
+{
+    public class PanoramicRoofState
+    {
+        public PanoramicRoofState()
+        {
+        }
+    }
+}
