+++ TeslaLib/Models/PerformanceConfiguration.cs
@@ -1,0 +1,17 @@
+﻿using System.Runtime.Serialization;
+
+namespace TeslaLib.Models
+{
+    public enum PerformanceConfiguration
+    {
+        [EnumMember(Value = "Base")]
+        BASE,
+
+        [EnumMember(Value = "Sport")]
+        SPORT,
+
+        [EnumMember(Value = "P2")]
+        P2
+    }}
+    }
+}
