+++ TeslaLib/Models/Region.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace TeslaLib.Models
+{
+    public class Region
+    {
+        public Region()
+        {
+        }
+    }
+}
