+++ TeslaLib/Models/TemperatureUnits.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.Serialization;
+
+namespace TeslaLib.Models
+{
+    public enum TemperatureUnits
+    {
+        [EnumMember(Value = "F")]
+        FAHRENHEIT,
+
+        [EnumMember(Value = "C")]
+        CELSIUS,
+    }
+}
